Extract functions to separate files.
